--- backend/src/main/java/shop/growith/api/Friends/domain/GestureCategoryDTO.java
@@ -1,14 +1,0 @@
-package shop.growith.api.Friends.domain;
-
-import lombok.Getter;
-import lombok.Setter;
-
-@Getter
-@Setter
-public class GestureCategoryDTO {
-
-    private Long id;
-
-    private String gestureCategory;
-}
-
