--- backend/src/main/java/shop/growith/api/Friends/service/FriendsServiceImpl.java
@@ -1,8 +1,0 @@
-package shop.growith.api.Friends.service;
-
-import org.springframework.stereotype.Service;
-
-@Service
-public class FriendsServiceImpl implements FriendsService{
-
-}
