+++ backend/src/main/java/shop/growith/api/friends/domain/FriendsListDTO.java
@@ -1,0 +1,18 @@
+package shop.growith.api.friends.domain;
+
+import lombok.Getter;
+import lombok.Setter;
+
+
+@Getter
+@Setter
+public class FriendsListDTO {
+
+    private long friendId;
+
+    private String userName;
+    private String nickname;
+    private Boolean favorite;
+    private String activityTime;
+
+}
