+++ backend/src/main/java/shop/growith/api/gesture/domain/GestureCategory.java
@@ -1,0 +1,23 @@
+package shop.growith.api.gesture.domain;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+import javax.persistence.*;
+
+@Entity
+@Getter
+@NoArgsConstructor
+@AllArgsConstructor
+@Builder
+@Table(name = "gesture_category")
+public class GestureCategory {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    @Column
+    private Long id; // 제스쳐 id
+
+    @Column    private String gestureCategory; // 제스쳐 카테고리
+}
