+++ backend/src/main/java/shop/growith/api/gesture/domain/GestureCategoryDTO.java
@@ -1,0 +1,14 @@
+package shop.growith.api.gesture.domain;
+
+import lombok.Getter;
+import lombok.Setter;
+
+@Getter
+@Setter
+public class GestureCategoryDTO {
+
+    private Long id;
+
+    private String gestureCategory;
+}
+
