+++ backend/src/main/java/shop/growith/api/gesture/service/GestureCatogory.java
@@ -1,0 +1,7 @@
+package shop.growith.api.gesture.service;
+
+public interface GestureCatogory {
+    void addGestureCategory(); // 제스쳐 카테고리 추가하기
+
+    void deleteGestureCategory(); // 제스쳐 카테고리 삭제하기
+}
